Fix : Fixed a sub-virion infection was not properly performed
There was a problem of saving all virus-infections.json to the main dir, so it was impossible to re-inject the virion that the sub virion depends on from the main.

Now when infecting a sub-virion, the working directory is also moved to the corresponding virion path.

To do this, I made a static variable that checks the deep of recursions, and the problem is solved :)